+++ rdf-abac-core/src/test/java/io/telicent/jena/abac/assembler/TS_ABAC_Assembler.java
@@ -24,6 +24,7 @@ import org.junit.platform.suite.api.Suite;
  */
 @Suite
 @SelectClasses({
-    TestAttributeStoreBuildLibCoverage.class
+    TestAttributeStoreBuildLibCoverage.class,
+    TestLabelStoreAssemblerModern.class
 })
 public class TS_ABAC_Assembler {}